types(frames): add types for frames
