+++ src/types/frames.ts
@@ -1,0 +1,9 @@
+export type Frame = {
+	snapshot: string;
+};
+
+export type FramesMap = Record<string, Frame>;
+
+export type Frames = {
+	frames: FramesMap;
+};
